refactor(ui): extract a shared PatchBlock for diagnosis patches
The suggested-patch block (validation badge, download, git apply, copy, diff
and the not-applicable error line) was implemented twice. Extract it into one
PatchBlock component and render the diagnosis result's patch through it.
